quic: move quic helpers from http/common into quic
