+++ authority/nonvoting/client/client.go
@@ -31,7 +31,7 @@ import (
 	"github.com/katzenpost/katzenpost/core/pki"
 	"github.com/katzenpost/katzenpost/core/wire"
 	"github.com/katzenpost/katzenpost/core/wire/commands"
-	"github.com/katzenpost/katzenpost/http/common"
+	"github.com/katzenpost/katzenpost/quic"
 	"gopkg.in/op/go-logging.v1"
 )
 
@@ -206,7 +206,7 @@ func (c *client) initSession(ctx context.Context, doneCh <-chan interface{}, sig
 	if err != nil {
 		return nil, nil, err
 	}
-	conn, err := common.DialURL(u, ctx, dialFn)
+	conn, err := quic.DialURL(u, ctx, dialFn)
 	if err != nil {
 		return nil, nil, err
 	}
