+++ quic/quic.go
@@ -13,9 +13,7 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-// Package common contains things shared by client and server
-package common
+package quic
 
 import (
 	"context"
